Move lyric base entity to separate module
