+++ homeassistant/components/lyric/sensor.py
@@ -25,7 +25,6 @@ from homeassistant.helpers.typing import StateType
 from homeassistant.helpers.update_coordinator import DataUpdateCoordinator
 from homeassistant.util import dt as dt_util
 
-from . import LyricAccessoryEntity, LyricDeviceEntity
 from .const import (
     DOMAIN,
     PRESET_HOLD_UNTIL,
@@ -34,6 +33,7 @@ from .const import (
     PRESET_TEMPORARY_HOLD,
     PRESET_VACATION_HOLD,
 )
+from .entity import LyricAccessoryEntity, LyricDeviceEntity
 
 LYRIC_SETPOINT_STATUS_NAMES = {
     PRESET_NO_HOLD: "Following Schedule",
